feat e refactor: readequação da página de gerenciarVendas para os pedidos pós venda
