+++ src/main/webapp/paginas/admin/gerenciarVendas/gerenciarVendas.html
@@ -34,7 +34,7 @@
                     <th>Operações</th>
                 </tr>
             </thead>
-            <tbody id="orderTable"></tbody>
+            <tbody id="tabela-pedidos"></tbody>
         </table>
 
         <h2 class="text-center mb-4">Pedidos de troca</h2>
@@ -43,6 +43,7 @@
                 <tr>
                     <th>#</th>
                     <th>Pedido</th>
+                    <th>Data</th>
                     <th>Cliente</th>
                     <th>Quadrinho</th>
                     <th>Status</th>
